correção exercicios

+++ exercicios/exercicio007.py
@@ -1,5 +1,5 @@
 nota1 = float(input('Qual é a primeira nota? '))
 nota2 = float(input('Qual é a segunda nota? '))
-total = float(nota1 + nota2)
-media = float(total/2)
+total = nota1 + nota2
+media = total/2
 print(f'A soma das duas notas é {total} e a média é {media:.1f}!')
